# Change [perf.sh] to use [dune-bench] repo by default