I'm switching to `https://github.com/ocaml-dune/dune-bench` as this is a better benchmark. 

I've also tweaked the script a bit so that it doesn't choke on build warnings and adds the repository name to the summary:

```
Summary for building https://github.com/ocaml-dune/dune-bench:

             Current branch  Main branch 
             ============== =============
           | real 54.138s   real 54.234s
Full build | user 168.245s  user 169.472s
           | sys  38.503s   sys  38.577s
                            
           | real 2.198s    real 2.197s
Zero build | user 2.046s    user 2.034s
           | sys  0.153s    sys  0.162s
```